Revert "license: Drop adding RRECOMMENDS for license packages"
This reverts commit 5348ffce46d6706b7bb10e41a59e0f6cf16c62b0.
To meet licensing requirements, *-lic packages should still
be added to RRECOMMENDS for PACKAGES in a recipe when the
LICENSE_CREATE_PACKAGE option is enabled. This was removed
in upstream as it did not apply to PACKAGES_DYNAMIC and
replaced with an option to add all licenses to an image
which was different from our use case (installing from
feeds).

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>