create-spdx: Skip missing files when building from feeds.
Currently, the methods for creating spdx sboms will fail
if an sbom for a package in the image is not found at
the expected path (tmp/deploy/spdx/<arch>/packages/).
In the case of building from feeds where the spdx files
will not be found locally, this is undesireable.

This modifies the behavior to only warn if a file is
missing when BUILD_IMAGES_FROM_FEEDS is enabled. This
is intended as a temporary solution until using other
locations for spdx files can be supported.

Signed-off-by: Charlie Johnston <charlie.johnston@ni.com>

Rebased-by: Alex Stewart <alex.stewart@ni.com>

Upstream has moved the create-spdx class logic to
create-spdx-2.2.bbclass. Rebased this change to that newer file.

(cherry picked from commit e236338f4ab626123251ed70cc2712c71de9c091)
Resolved merge conflicts by retaining our changes.
Signed-off-by: Rajendra Desai <rajendra.desai@emerson.com>