spdx30: Skip SPDX processing for uncontrolled packages
When building images from feeds, some packages are not controlled by
OpenEmbedded and do not provide SPDX 3.0 artifacts. This change
updates the SPDX 3.0 tasks to skip processing for such packages to
avoid build failures and unnecessary warnings.

This aligns with prior behavior implemented for SPDX 2.2, but applies
the logic to SPDX 3.0 now that it is the default version.

Signed-off-by: Shubhanshu Mani Tripathi <shubhanshu.mani.tripathi@emerson.com>